# Fix performance of merging Dicts by sizehint!