@petvana 2022-03-18
Not sure if news is needed here. I'd think this is a kind of bugfix with no new functionality. The only change is that it more likely triggers resizing.